some cleanup with path_actions module
